Use type long for delay as it corresp[onds with the alsa type used

+++ audio.h
@@ -32,7 +32,7 @@ typedef struct {
   // almost certainly wrong if the buffer is empty, so put silent buffers into it to make it busy.
   // will change dynamically, so keep watching it. Implemented in ALSA only.
   // returns -1 if there's a problem
-  int32_t (*delay)();
+  long (*delay)(); // snd_pcm_sframes_t is a signed long
 
   // may be NULL, in which case soft volume is applied
   void (*volume)(double vol);
